# static/swagger.yml
openapi: 3.0.3
info:
  description: "API для КИП"
  version: "1.0.0"
  title: "КИП API"
host: "localhost:5000"
basePath: "/"
...
schemes:
  - "http"
paths:
  /universal/{table_name}:
    post:
      tags:
        - "all"
      summary: "Добавить данные в указаную таблицу"
      parameters:
        - name: "table_name"
          in: "path"
          required: true
          description: "Таблица, куда добавить данные"
          type: "string"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
      responses:
        200:
          description: "Добавлено в таблицу"
        400:
          description: "He добавлено"
    get:
      tags:
        - "all"
      summary: "Получить данные из указаной таблицы"
      parameters:
        - name: "table_name"
          in: "path"
          required: true
          description: "Таблица, откуда брать данные"
          type: "string"
        - name: "filters"
          in: "query"
          description: "Условие"
          required: false
          schema:
            type: object
            additionalProperties:
              type: string
          style: form
          explode: true
      responses:
        200:
          description: "Получено"
        400:
          description: "He получено"
    put:
      tags:
        - "all"
      summary: "Изменить данные в указаной таблицы"
      parameters:
        - name: "table_name"
          in: "path"
          required: true
          description: "Таблица, откуда брать данные"
          type: "string"
        - name: "filters"
          in: "query"
          description: "Условие"
          required: true
          schema:
            type: object
            additionalProperties:
              type: string
          style: form
          explode: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
      responses:
        200:
          description: "изменено"
        400:
          description: "He изменено"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...